refactor(live-trade): extract order audit services
