+++ src/gpt_trader/features/live_trade/engines/strategy.py
@@ -17,8 +17,7 @@ import sys
 import threading
 import time
 from collections import deque
-from dataclasses import replace
-from datetime import datetime, timezone
+from datetime import datetime
 from decimal import Decimal
 from importlib import metadata
 from typing import Any
@@ -38,6 +37,10 @@ from gpt_trader.features.live_trade.engines.base import (
 )
 from gpt_trader.features.live_trade.engines.equity_calculator import (
     EquityCalculator,
+)
+from gpt_trader.features.live_trade.engines.order_audit import OrderAuditService
+from gpt_trader.features.live_trade.engines.order_reconciliation import (
+    OrderReconciliationService,
 )
 from gpt_trader.features.live_trade.engines.order_record_mapping import (
     build_record_from_broker_order,
@@ -83,7 +86,6 @@ from gpt_trader.features.live_trade.engines.telemetry_streaming import (
     _start_streaming,
     _stop_streaming,
 )
-from gpt_trader.features.live_trade.execution.broker_executor import BrokerExecutor
 from gpt_trader.features.live_trade.execution.decision_trace import OrderDecisionTrace
 from gpt_trader.features.live_trade.execution.guard_manager import GuardManager
 from gpt_trader.features.live_trade.execution.order_submission import OrderSubmitter
@@ -193,9 +195,6 @@ class TradingEngine(BaseEngine):
 
         # Initialize graceful degradation state
         self._degradation = DegradationState()
-        self._unfilled_order_alerts: dict[str, float] = {}
-        self._order_reconciliation_drift_failures = 0
-        self._order_reconciliation_drift_escalated = False
 
         broker_calls = getattr(context, "broker_calls", None)
         if broker_calls is not None and not asyncio.iscoroutinefunction(
@@ -328,6 +327,25 @@ class TradingEngine(BaseEngine):
             enable_retries=getattr(self.context.config, "order_submission_retries_enabled", False),
             orders_store=self._orders_store,
             integration_mode=False,
+        )
+        self._order_reconciliation = OrderReconciliationService(
+            context=self.context,
+            open_orders=self._open_orders,
+            orders_store_provider=lambda: self._orders_store,
+            broker_calls_provider=lambda: self._broker_calls,
+            degradation=self._degradation,
+            order_submitter=self._order_submitter,
+            append_event=self._append_event,
+            notify=self._notify,
+            drift_max_failures=ORDER_RECONCILIATION_DRIFT_MAX_FAILURES,
+        )
+        self._order_audit = OrderAuditService(
+            context=self.context,
+            broker_calls_provider=lambda: self._broker_calls,
+            status_reporter=self._status_reporter,
+            reconciliation=self._order_reconciliation,
+            append_event=self._append_event,
+            cycle_count_provider=lambda: self._cycle_count,
         )
 
         # Failure tracker from container (not global) with escalation callback
@@ -469,51 +487,10 @@ class TradingEngine(BaseEngine):
         *,
         bot_id: str,
     ) -> list[OrderRecord]:
-        if not orders or self._orders_store is None:
-            return []
-        now = datetime.now(timezone.utc)
-        recovered: list[OrderRecord] = []
-        for order in orders:
-            record = self._build_record_from_broker_order(order, bot_id=bot_id, now=now)
-            if record is None:
-                continue
-            metadata = dict(record.metadata or {})
-            metadata["note"] = "backfilled_from_broker"
-            record = replace(record, metadata=metadata)
-            try:
-                await self._broker_calls(self._orders_store.upsert_by_client_id, record)
-            except Exception as exc:
-                logger.warning(
-                    "Failed to persist recovered order",
-                    error_type=type(exc).__name__,
-                    error_message=str(exc),
-                    operation="order_reconciliation",
-                    stage="persist_recovery",
-                    order_id=record.order_id,
-                    client_order_id=record.client_order_id,
-                )
-                continue
-            if record.order_id not in self._open_orders:
-                self._open_orders.append(record.order_id)
-            recovered.append(record)
-
-        if recovered:
-            payload = {
-                "timestamp": time.time(),
-                "bot_id": bot_id,
-                "recovered_order_ids": [record.order_id for record in recovered][:10],
-                "recovered_order_count": len(recovered),
-                "open_order_count": len(self._open_orders),
-            }
-            self._append_event("order_reconciliation_recovered", payload)
-            logger.info(
-                "Recovered unknown bot orders",
-                recovered_order_count=len(recovered),
-                operation="order_reconciliation",
-                stage="recover",
-            )
-
-        return recovered
+        return await self._order_reconciliation.recover_unknown_bot_orders(
+            orders,
+            bot_id=bot_id,
+        )
 
     async def _refresh_missing_persisted_orders(
         self,
@@ -521,92 +498,10 @@ class TradingEngine(BaseEngine):
         *,
         bot_id: str,
     ) -> list[Any]:
-        if not records or self._orders_store is None:
-            return []
-        broker = self.context.broker
-        if broker is None:
-            return []
-        get_order = getattr(broker, "get_order", None)
-        if not callable(get_order):
-            return []
-        now = datetime.now(timezone.utc)
-        refreshed_orders: list[Any] = []
-        for record in records:
-            try:
-                order = await self._broker_calls(get_order, record.order_id)
-                if order is None and record.client_order_id != record.order_id:
-                    order = await self._broker_calls(get_order, record.client_order_id)
-            except Exception as exc:
-                logger.warning(
-                    "Failed to load order during reconciliation",
-                    error_type=type(exc).__name__,
-                    error_message=str(exc),
-                    operation="order_reconciliation",
-                    stage="load_missing",
-                    order_id=record.order_id,
-                    client_order_id=record.client_order_id,
-                )
-                continue
-            if order is None:
-                continue
-            update_record = self._build_record_from_broker_order(order, bot_id=bot_id, now=now)
-            if update_record is None:
-                continue
-            update_metadata = dict(update_record.metadata or {})
-            update_metadata["note"] = "refresh_missing"
-            metadata = self._merge_metadata(record.metadata, update_metadata)
-            updated = replace(
-                record,
-                order_id=update_record.order_id,
-                client_order_id=update_record.client_order_id,
-                symbol=update_record.symbol or record.symbol,
-                side=update_record.side or record.side,
-                order_type=update_record.order_type or record.order_type,
-                quantity=(
-                    update_record.quantity
-                    if update_record.quantity != Decimal("0")
-                    else record.quantity
-                ),
-                price=update_record.price if update_record.price is not None else record.price,
-                status=update_record.status,
-                filled_quantity=(
-                    update_record.filled_quantity
-                    if update_record.filled_quantity != Decimal("0")
-                    else record.filled_quantity
-                ),
-                average_fill_price=(update_record.average_fill_price or record.average_fill_price),
-                updated_at=now,
-                metadata=metadata,
-                bot_id=record.bot_id or bot_id,
-                time_in_force=update_record.time_in_force or record.time_in_force,
-            )
-            try:
-                await self._broker_calls(self._orders_store.upsert_by_client_id, updated)
-            except Exception as exc:
-                logger.warning(
-                    "Failed to persist order refresh",
-                    error_type=type(exc).__name__,
-                    error_message=str(exc),
-                    operation="order_reconciliation",
-                    stage="persist_refresh",
-                    order_id=updated.order_id,
-                    client_order_id=updated.client_order_id,
-                )
-                continue
-
-            if updated.status in {
-                PersistedOrderStatus.PENDING,
-                PersistedOrderStatus.OPEN,
-                PersistedOrderStatus.PARTIALLY_FILLED,
-            }:
-                if updated.order_id not in self._open_orders:
-                    self._open_orders.append(updated.order_id)
-                refreshed_orders.append(order)
-            else:
-                if updated.order_id in self._open_orders:
-                    self._open_orders.remove(updated.order_id)
-
-        return refreshed_orders
+        return await self._order_reconciliation.refresh_missing_persisted_orders(
+            records,
+            bot_id=bot_id,
+        )
 
     async def _reconcile_open_orders(self, orders: list[Any]) -> None:
         """Reconcile internal open-order tracking with broker + persistence state.
@@ -617,282 +512,11 @@ class TradingEngine(BaseEngine):
         submit_id into ``self._open_orders``. On the next audit, we normalize tracked
         IDs back to broker order IDs when the broker returns both IDs.
         """
-        bot_id = str(self.context.bot_id or self.context.config.profile or "live")
-        prefix = f"{bot_id}_"
-
-        pending: list[OrderRecord] = []
-        pending_by_client_id: dict[str, OrderRecord] = {}
-        tracked_ids: set[str] = {str(order_id) for order_id in self._open_orders}
-
-        if self._orders_store is not None:
-            try:
-                pending = await self._broker_calls(
-                    self._orders_store.get_pending_orders,
-                    bot_id=bot_id or None,
-                )
-            except Exception as exc:
-                logger.warning(
-                    "Failed to load pending orders during audit",
-                    error_type=type(exc).__name__,
-                    error_message=str(exc),
-                    operation="order_reconciliation",
-                    stage="load_pending",
-                )
-                pending = []
-
-        for store_record in pending:
-            pending_by_client_id[str(store_record.client_order_id)] = store_record
-            tracked_ids.add(str(store_record.client_order_id))
-            tracked_ids.add(str(store_record.order_id))
-
-        if not orders:
-            refreshed_orders: list[Any] = []
-            if pending:
-                refreshed_orders = await self._refresh_missing_persisted_orders(
-                    pending,
-                    bot_id=bot_id,
-                )
-            if refreshed_orders:
-                orders = refreshed_orders
-            else:
-                if not pending and self._open_orders:
-                    self._open_orders.clear()
-                self._order_reconciliation_drift_failures = 0
-                self._order_reconciliation_drift_escalated = False
-                return
-
-        order_index: dict[str, Any] = {}
-
-        def _index_order(order: Any) -> None:
-            if order is None:
-                return
-            order_id = self._get_order_field(order, "order_id", "id")
-            client_order_id = self._get_order_field(order, "client_order_id", "client_id")
-            if order_id is not None:
-                order_index[str(order_id)] = order
-            if client_order_id is not None:
-                order_index[str(client_order_id)] = order
-
-        for order in orders:
-            _index_order(order)
-
-        # Update persisted records that are still using submit_id as order_id, now that
-        # the broker has returned the canonical order_id.
-        if self._orders_store is not None and pending_by_client_id:
-            now = datetime.now(timezone.utc)
-            for order in orders:
-                order_id = self._get_order_field(order, "order_id", "id")
-                client_order_id = self._get_order_field(order, "client_order_id", "client_id")
-                if order_id is None or client_order_id is None:
-                    continue
-                client_id_str = str(client_order_id)
-                pending_record = pending_by_client_id.get(client_id_str)
-                if pending_record is None:
-                    continue
-                if pending_record.order_id != pending_record.client_order_id:
-                    continue
-                order_id_str = str(order_id)
-                if order_id_str == pending_record.order_id:
-                    continue
-                try:
-                    updated = replace(
-                        pending_record,
-                        order_id=order_id_str,
-                        status=PersistedOrderStatus.OPEN,
-                        updated_at=now,
-                        metadata={
-                            **(pending_record.metadata or {}),
-                            "source": "order_audit",
-                            "note": "normalized_submit_id_to_order_id",
-                        },
-                    )
-                    await self._broker_calls(self._orders_store.upsert_by_client_id, updated)
-                except Exception as exc:
-                    logger.warning(
-                        "Failed to persist order_id normalization",
-                        error_type=type(exc).__name__,
-                        error_message=str(exc),
-                        operation="order_reconciliation",
-                        stage="persist_normalization",
-                        client_order_id=client_id_str,
-                        order_id=order_id_str,
-                    )
-
-        if pending:
-            open_ids = set(order_index.keys())
-            missing_records = [
-                record
-                for record in pending
-                if str(record.order_id) not in open_ids
-                and str(record.client_order_id) not in open_ids
-            ]
-            refreshed_orders = await self._refresh_missing_persisted_orders(
-                missing_records,
-                bot_id=bot_id,
-            )
-            for refreshed_order in refreshed_orders:
-                _index_order(refreshed_order)
-
-        # Reconcile the tracked IDs: replace submit_ids with canonical order_ids and drop
-        # IDs that are no longer open at the broker.
-        snapshot = list(self._open_orders)
-        snapshot_set = set(snapshot)
-        reconciled: list[str] = []
-        seen: set[str] = set()
-
-        def _canonical_order_id(tracked_id: str) -> str | None:
-            order = order_index.get(tracked_id)
-            if order is None:
-                return None
-            oid = self._get_order_field(order, "order_id", "id") or tracked_id
-            return str(oid)
-
-        for tracked_id in snapshot:
-            canonical = _canonical_order_id(tracked_id)
-            if canonical is None or canonical in seen:
-                continue
-            seen.add(canonical)
-            reconciled.append(canonical)
-
-        # Preserve any IDs appended while we were reconciling (e.g., order submission)
-        # without losing them via slice assignment.
-        for tracked_id in list(self._open_orders):
-            if tracked_id in snapshot_set:
-                continue
-            canonical = _canonical_order_id(tracked_id) or str(tracked_id)
-            if canonical in seen:
-                continue
-            seen.add(canonical)
-            reconciled.append(canonical)
-
-        if reconciled != snapshot:
-            self._open_orders[:] = reconciled
-
-        # Drift detection: bot-owned open orders present at the broker but not tracked in
-        # the store/open-orders list can indicate "ghost" orders after a crash.
-        unknown_bot_orders: list[str] = []
-        unknown_orders: list[Any] = []
-        for order in orders:
-            client_order_id = self._get_order_field(order, "client_order_id", "client_id")
-            if client_order_id is None:
-                continue
-            client_id_str = str(client_order_id)
-            if not client_id_str.startswith(prefix):
-                continue
-            order_id = self._get_order_field(order, "order_id", "id")
-            order_id_str = str(order_id) if order_id is not None else client_id_str
-            if order_id_str in tracked_ids or client_id_str in tracked_ids:
-                continue
-            unknown_bot_orders.append(order_id_str)
-            unknown_orders.append(order)
-
-        recovered_records = await self._recover_unknown_bot_orders(
-            unknown_orders,
-            bot_id=bot_id,
-        )
-        if recovered_records:
-            recovered_ids = {record.order_id for record in recovered_records}
-            recovered_ids.update(record.client_order_id for record in recovered_records)
-            tracked_ids.update(recovered_ids)
-            unknown_bot_orders = [
-                order_id for order_id in unknown_bot_orders if order_id not in recovered_ids
-            ]
-
-        if not unknown_bot_orders:
-            self._order_reconciliation_drift_failures = 0
-            self._order_reconciliation_drift_escalated = False
-            return
-
-        # Record drift event for observability.
-        payload = {
-            "timestamp": time.time(),
-            "bot_id": bot_id,
-            "unknown_bot_order_ids": unknown_bot_orders[:10],
-            "unknown_bot_order_count": len(unknown_bot_orders),
-            "open_order_count": len(orders),
-        }
-        self._append_event("order_reconciliation_drift", payload)
-        logger.warning(
-            "Order reconciliation drift detected",
-            unknown_bot_order_count=len(unknown_bot_orders),
-            open_order_count=len(orders),
-            operation="order_reconciliation",
-            stage="drift",
-        )
-
-        if self._order_reconciliation_drift_escalated:
-            return
-
-        self._order_reconciliation_drift_failures += 1
-        if self._order_reconciliation_drift_failures < ORDER_RECONCILIATION_DRIFT_MAX_FAILURES:
-            return
-
-        self._order_reconciliation_drift_escalated = True
-        await self._handle_order_reconciliation_drift(unknown_bot_orders)
+        await self._order_reconciliation.reconcile_open_orders(orders)
 
     async def _handle_order_reconciliation_drift(self, order_ids: list[str]) -> None:
         """Escalate persistent order-reconciliation drift to graceful degradation."""
-        risk_manager = self.context.risk_manager
-        config = risk_manager.config if risk_manager else None
-
-        cooldown_seconds = getattr(config, "api_health_cooldown_seconds", 300) if config else 300
-        if risk_manager is not None:
-            risk_manager.set_reduce_only_mode(True, reason="order_reconciliation_drift")
-
-        broker = self.context.broker
-        if broker is not None:
-            cancel_executor = BrokerExecutor(broker=broker)
-            retry_enabled = getattr(self.context.config, "order_submission_retries_enabled", False)
-            for order_id in order_ids:
-                try:
-                    await self._broker_calls(
-                        cancel_executor.cancel_order,
-                        order_id,
-                        use_retry=retry_enabled,
-                        allow_idempotent=True,
-                    )
-                except Exception as exc:
-                    logger.warning(
-                        "Failed to cancel drifted order",
-                        error_type=type(exc).__name__,
-                        error_message=str(exc),
-                        operation="order_reconciliation",
-                        stage="cancel",
-                        order_id=order_id,
-                    )
-
-        self._degradation.pause_all(
-            seconds=cooldown_seconds,
-            reason="order_reconciliation_drift",
-            allow_reduce_only=True,
-        )
-
-        # Record a synthetic rejection for telemetry.
-        try:
-            self._order_submitter.record_rejection(
-                symbol="*",
-                side="*",
-                quantity=Decimal("0"),
-                price=None,
-                reason="order_reconciliation_drift",
-            )
-        except Exception:
-            pass
-
-        await self._notify(
-            title="Order Reconciliation Drift - Trading Paused",
-            message=(
-                "Detected open broker orders that appear to belong to this bot but are not tracked. "
-                f"Paused trading for {cooldown_seconds}s and enabled reduce-only mode. "
-                f"Attempted cancellation for {len(order_ids)} order(s)."
-            ),
-            severity=AlertSeverity.ERROR,
-            context={
-                "unknown_order_ids": order_ids[:10],
-                "unknown_order_count": len(order_ids),
-                "cooldown_seconds": cooldown_seconds,
-            },
-        )
+        await self._order_reconciliation.handle_order_reconciliation_drift(order_ids)
 
     # =========================================================================
     # Streaming Lifecycle Methods
@@ -2356,58 +1980,7 @@ class TradingEngine(BaseEngine):
         return parse_timestamp(value)
 
     def _record_unfilled_order_alerts(self, orders: list[Any]) -> None:
-        risk_manager = self.context.risk_manager
-        config = risk_manager.config if risk_manager else None
-        threshold = getattr(config, "unfilled_order_alert_seconds", 300)
-        if not orders or threshold <= 0:
-            self._unfilled_order_alerts.clear()
-            return
-
-        now = time.time()
-        active_ids: set[str] = set()
-
-        for order in orders:
-            order_id = self._get_order_field(
-                order, "order_id", "id", "client_order_id", "client_id"
-            )
-            if order_id is None:
-                continue
-            order_id_str = str(order_id)
-            active_ids.add(order_id_str)
-            if order_id_str in self._unfilled_order_alerts:
-                continue
-
-            status = self._get_order_field(order, "status")
-            status_str = (status.value if hasattr(status, "value") else str(status or "")).upper()
-            if status_str in {"FILLED", "CANCELLED", "CANCELED", "REJECTED"}:
-                continue
-
-            created_at = self._get_order_field(
-                order, "created_time", "created_at", "submitted_at", "created"
-            )
-            created_ts = self._parse_timestamp(created_at)
-            age_seconds = now - created_ts
-            if age_seconds < threshold:
-                continue
-
-            payload = {
-                "order_id": order_id_str,
-                "symbol": self._get_order_field(order, "product_id", "symbol") or "",
-                "side": self._get_order_field(order, "side") or "",
-                "status": status_str,
-                "created_time": created_ts,
-                "age_seconds": age_seconds,
-                "threshold_seconds": threshold,
-                "timestamp": now,
-            }
-            self._append_event("unfilled_order_alert", payload)
-            self._unfilled_order_alerts[order_id_str] = now
-
-        stale_ids = [
-            order_id for order_id in self._unfilled_order_alerts if order_id not in active_ids
-        ]
-        for order_id in stale_ids:
-            del self._unfilled_order_alerts[order_id]
+        self._order_audit.record_unfilled_order_alerts(orders)
 
     def _finalize_decision_trace(
         self,
@@ -3448,91 +3021,4 @@ class TradingEngine(BaseEngine):
 
     async def _audit_orders(self) -> None:
         """Audit open orders for reconciliation."""
-        broker = self.context.broker
-        if broker is None:
-            return
-        try:
-            # Fetch open orders
-            response: Any = None
-            module_name = getattr(broker, "__module__", "")
-            if "coinbase" in module_name:
-                # Prefer Coinbase client list_orders to get the raw dict shape expected by StatusReporter.
-                client = getattr(broker, "client", None)
-                client_list_orders = getattr(client, "list_orders", None)
-                if callable(client_list_orders):
-                    response = await self._broker_calls(client_list_orders, order_status="OPEN")
-            if response is None:
-                list_orders = getattr(broker, "list_orders", None)
-                if callable(list_orders):
-                    try:
-                        response = await self._broker_calls(list_orders, order_status="OPEN")
-                    except TypeError:
-                        response = await self._broker_calls(list_orders, status=["OPEN"])
-
-            orders: list[Any]
-            if isinstance(response, dict):
-                orders = list(response.get("orders", []))
-            elif isinstance(response, list):
-                orders = list(response)
-            else:
-                orders = []
-
-            await self._reconcile_open_orders(orders)
-
-            if orders:
-                logger.info(
-                    "AUDIT: Found OPEN orders",
-                    open_order_count=len(orders),
-                    operation="order_audit",
-                    stage="list",
-                )
-
-            self._record_unfilled_order_alerts(orders)
-
-            # Update status reporter (expects dict-like orders).
-            status_orders: list[dict[str, Any]] = []
-            for order in orders:
-                if isinstance(order, dict):
-                    status_orders.append(order)
-                    continue
-                status_orders.append(
-                    {
-                        "order_id": self._get_order_field(order, "order_id", "id") or "",
-                        "product_id": self._get_order_field(order, "product_id", "symbol") or "",
-                        "side": self._get_order_field(order, "side") or "",
-                        "status": self._get_order_field(order, "status") or "",
-                        "price": self._get_order_field(order, "price"),
-                        "size": self._get_order_field(order, "size", "quantity"),
-                        "created_time": self._get_order_field(order, "created_time", "created_at"),
-                        "filled_size": self._get_order_field(
-                            order, "filled_size", "filled_quantity"
-                        ),
-                        "average_filled_price": self._get_order_field(
-                            order, "average_filled_price", "avg_fill_price"
-                        ),
-                    }
-                )
-            self._status_reporter.update_orders(status_orders)
-
-            # Update Account Metrics (every 60 cycles ~ 1 minute)
-            if self._cycle_count % 60 == 0:
-                try:
-                    balances = await self._broker_calls(broker.list_balances)
-                    # Check if broker supports transaction summary (Coinbase specific)
-                    summary = {}
-                    if hasattr(broker, "client") and hasattr(
-                        broker.client, "get_transaction_summary"
-                    ):
-                        try:
-                            summary = await self._broker_calls(
-                                broker.client.get_transaction_summary
-                            )
-                        except Exception:
-                            pass  # Feature might not be available or API mode issue
-
-                    self._status_reporter.update_account(balances, summary)
-                except Exception as e:
-                    logger.warning(f"Failed to update account metrics: {e}")
-
-        except Exception as e:
-            logger.warning(f"Failed to audit orders: {e}")
+        await self._order_audit.audit_orders()
