+++ var/agents/logging/event_catalog.json
@@ -1331,6 +1331,7 @@
       ],
       "occurrence_count": 2,
       "source_files": [
+        "features/live_trade/engines/order_audit.py",
         "features/live_trade/engines/strategy.py"
       ],
       "statuses": null
@@ -1430,7 +1431,7 @@
       ],
       "occurrence_count": 8,
       "source_files": [
-        "features/live_trade/engines/strategy.py"
+        "features/live_trade/engines/order_reconciliation.py"
       ],
       "statuses": null
     },
