+++ var/agents/observability/metrics_catalog.md
@@ -10,7 +10,7 @@ Generated by `scripts/agents/generate_metrics_catalog.py`.
 | `gpt_trader_api_retries_total` | counter | category, reason | src/gpt_trader/features/brokerages/coinbase/client/base.py:440 |
 | `gpt_trader_broker_call_latency_seconds` | histogram | operation, outcome, reason | src/gpt_trader/features/live_trade/execution/broker_executor.py:95 |
 | `gpt_trader_circuit_breaker_state` | gauge | category | src/gpt_trader/features/brokerages/coinbase/client/base.py:694 |
-| `gpt_trader_cycle_duration_seconds` | histogram | result | src/gpt_trader/features/live_trade/engines/strategy.py:1664 |
+| `gpt_trader_cycle_duration_seconds` | histogram | result | src/gpt_trader/features/live_trade/engines/strategy.py:1288 |
 | `gpt_trader_data_quality_issues` | gauge | interval, symbol | src/gpt_trader/features/data/data.py:281 |
 | `gpt_trader_data_quality_issues_total` | counter | interval, issue_type, severity, symbol | src/gpt_trader/features/data/data.py:288 |
 | `gpt_trader_data_quality_score` | gauge | interval, symbol | src/gpt_trader/features/data/data.py:276 |
@@ -23,7 +23,7 @@ Generated by `scripts/agents/generate_metrics_catalog.py`.
 | `gpt_trader_order_missing_decision_id_total` | counter |  | src/gpt_trader/features/live_trade/execution/order_submission.py:665 |
 | `gpt_trader_order_submission_latency_seconds` | histogram | result, side | src/gpt_trader/features/live_trade/execution/order_submission.py:204 |
 | `gpt_trader_order_submission_total` | counter | reason, result, side | src/gpt_trader/features/live_trade/execution/order_submission.py:178 |
-| `gpt_trader_positions_fetch_seconds` | histogram | result | src/gpt_trader/features/live_trade/engines/strategy.py:2102 |
+| `gpt_trader_positions_fetch_seconds` | histogram | result | src/gpt_trader/features/live_trade/engines/strategy.py:1726 |
 | `gpt_trader_process_memory_mb` | gauge |  | src/gpt_trader/features/live_trade/engines/system_maintenance.py:141 |
 | `gpt_trader_profile_duration_seconds` | histogram |  | src/gpt_trader/monitoring/profiling.py:67 |
 | `gpt_trader_rate_limit_usage_ratio` | gauge |  | src/gpt_trader/features/brokerages/coinbase/client/base.py:681 |
